Move session modules into their own files
